Fix compilation error on arm32
PiperOrigin-RevId: 262628782

+++ tensorflow/lite/experimental/ruy/pmu.cc
@@ -49,7 +49,8 @@ class PerfEvent {
     pe.exclude_hv = 1;
     fd_ = syscall(__NR_perf_event_open, &pe, 0, -1, -1, 0);
     if (fd_ == -1) {
-      fprintf(stderr, "perf_event_open failed for config 0x%lx\n", config);
+      fprintf(stderr, "perf_event_open failed for config 0x%lx\n",
+              static_cast<unsigned long>(config));
       // abort();
     }
     ioctl(fd_, PERF_EVENT_IOC_RESET, 0);
